+++ app/core/views/main.py
@@ -33,13 +33,6 @@ def help(request):
 
     return render(request, "react.html", context)
 
-def profile(request):
-    context = {
-        "title": "Profile",
-        "js_resources": ["dist/js/profile.js"],
-        "css_resources": ["dist/css/profile.css"],
-    }
-    return render(request, "react.html", context)
 
 def handler404(request, exception):
     # Log the 404 URL
